+++ smarts/core/sumo_traffic_simulation.py
@@ -52,7 +52,15 @@ from smarts.core.utils.sumo import traci, TraciConn  # isort:skip
 import traci.constants as tc  # isort:skip
 
 
-class SumoTrafficSimulation(TrafficProvider):
+class Itermediary(TrafficProvider):
+    pass
+
+
+class Other:
+    pass
+
+
+class SumoTrafficSimulation(Itermediary, Other):
     """
     Args:
         headless: