+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Traffic lights are now rendered in camera observations.
 - Interest vehicles now show up in Envision.
 - Seed of `hiway-v1` env can be retrieved through a new property `seed`.
+- Added `TrafficEngineActor` to describe a scenario studio defined actor that is controlled by a traffic engine.
 ### Changed
 - Changed waypoints in sumo maps to use more incoming lanes into junctions.
 - Increased the cutoff radius for filtering out waypoints that are too far away in junctions in sumo maps.
@@ -54,6 +55,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Removed `intersection-v0` env and `examples/rl/intersection` as they are superseded by `driving-smarts-v2023` env and `examples/rl/drive`, respectively.
 - Removed `examples/rl/racing` as base repository `danijar/dreamerv2` is not updated anymore and this example will be superseded by `examples/rl/drive`.
 - Removed `FrameStack` environment wrapper.
+- Removed `SMARTS.traffic_sim` property.
 ### Security
 
 ## [1.1.0] # 2023-04-28